requirements.txt
binance-futures-connector==1.6.0
python-dotenv==1.0.0